requirements.txt
moviepy
deepgram-sdk
python-dotenv